* Code reorganization
