+++ runtime/solo/src/lib.rs
@@ -71,6 +71,7 @@ use sp_version::RuntimeVersion;
 mod consts;
 mod fee_handler;
 mod gov;
+mod sage;
 mod types;
 
 pub use crate::types::{
@@ -220,7 +221,6 @@ impl frame_system::Config for Runtime {
 
 type SingleBlockMigrations = ();
 
-use crate::fee_handler::{CasinoJamFeeHandler, DummyVoucherHandler, NativeAndAssets};
 #[cfg(not(feature = "runtime-benchmarks"))]
 use mbm::MultiBlockMigrations;
 
@@ -549,233 +549,6 @@ impl pallet_proxy::Config for Runtime {
 	type CallHasher = BlakeTwo256;
 	type AnnouncementDepositBase = AnnouncementDepositBase;
 	type AnnouncementDepositFactor = AnnouncementDepositFactor;
-}
-
-impl pallet_insecure_randomness_collective_flip::Config for Runtime {}
-
-pub struct SageEngine;
-
-/// Runtime specific sage implementation so that we don't have to
-/// pass our type definitions all the time.
-macro_rules! impl_runtime_sage_api {
-	(
-		$impl_target:ident,
-		$runtime:ident,
-		$sage_instance:ident,
-		$season_manager:ident,
-		$asset_id:ident,
-		$asset:ident,
-		$transition_config:ident,
-	) => {
-		impl_sage_api!(
-			$impl_target,
-			$runtime,
-			$sage_instance,
-			$season_manager,
-			CasinoJamRandom,
-			AccountId,
-			$asset_id,
-			$asset,
-			FungiblesAssetId,
-			Balance,
-			BlockNumber,
-			CasinoJamSeasonId,
-			$transition_config,
-			H256,
-		);
-	};
-}
-
-// Every new game we add can simply call that macro for another sage instance to
-// implement the sage api given that the other types are identical.
-impl_runtime_sage_api!(
-	SageEngine,
-	Runtime,
-	CasinoJamSageInstance,
-	CasinoJamSeasons,
-	AssetId,
-	CasinoJamAsset,
-	CasinoJamTransitionConfig,
-);
-
-parameter_types! {
-	pub const SageHeroJamId: PalletId = PalletId(*b"sage/hjm");
-}
-
-pub type CasinoJamAsset = Asset<BlockNumberFor<Runtime>>;
-pub type GameTransition = CasinoJamTransition<AccountId, BlockNumberFor<Runtime>, SageEngine>;
-
-pub type CasinoJamAssetFilter = GameFilter<BlockNumberFor<Runtime>>;
-pub type CasinoJamBenchmarkHelper = GameBenchmarkHelper<BlockNumberFor<Runtime>>;
-
-type FungiblesAssetId = WithdrawKind<NativeOrWithId<AssetId>>;
-type TransferWithdraw =
-	WithdrawCreditOrVoucher<WithdrawFungibles<AccountId, NativeAndAssets>, DummyVoucherHandler>;
-
-pub type CasinoJamSageInstance = pallet_sage::Instance1;
-impl pallet_sage::Config<CasinoJamSageInstance> for Runtime {
-	type PalletId = SageHeroJamId;
-	type SageGameTransition = GameTransition;
-	type SeasonHandler = CasinoJamSeasons;
-	type FeeHandler = CasinoJamFeeHandler;
-	type TransferFunds = TransferFungibleAssets<TransferWithdraw, FungiblesAssetId>;
-	type FungiblesAssetId = FungiblesAssetId;
-	type FilterHandler = CasinoJamAssetFilter;
-	type Fungible = Balances;
-	type RuntimeEvent = RuntimeEvent;
-	type WeightInfo = ();
-	#[cfg(feature = "runtime-benchmarks")]
-	type BenchmarkHelper = CasinoJamBenchmarkHelper;
-}
-
-pub type CasinoJamSeasonId = u32;
-
-pub type SeasonsHeroJamInstance = pallet_ajuna_seasons::Instance1;
-impl pallet_ajuna_seasons::Config<SeasonsHeroJamInstance> for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type SeasonId = CasinoJamSeasonId;
-	type AssetId = AssetId;
-	type AccountHandler = CasinoJamSage;
-	type Currency = Balances;
-	type WeightInfo = ();
-	#[cfg(feature = "runtime-benchmarks")]
-	type BenchmarkHelper = ();
-}
-
-parameter_types! {
-	pub const AffiliateMaxLevel: u32 = 2;
-}
-
-pub type HeroJamRuleIdentifier = AffiliateMethods<CasinoAction>;
-
-#[cfg(feature = "runtime-benchmarks")]
-pub struct AffiliatesHeroJamBenchmarkHelper;
-
-#[cfg(feature = "runtime-benchmarks")]
-impl pallet_ajuna_affiliates::BenchmarkHelper<HeroJamRuleIdentifier, ()>
-	for AffiliatesHeroJamBenchmarkHelper
-{
-	fn create_rule_id(_id: u32) -> HeroJamRuleIdentifier {
-		AffiliateMethods::StateTransition(CasinoAction::Create(AssetType::Player))
-	}
-
-	fn create_params(_id: u32) {}
-}
-
-pub struct MockAffiliateUnlockRules;
-
-impl AffiliateUnlockRules for MockAffiliateUnlockRules {
-	type AccountId = AccountId;
-	type UnlockParameters = ();
-
-	fn execute_unlock_rule_for(
-		_account: &Self::AccountId,
-		_params: Self::UnlockParameters,
-	) -> Result<(), DispatchError> {
-		Ok(())
-	}
-}
-
-pub type AffiliatesHeroJamInstance = pallet_ajuna_affiliates::Instance1;
-impl pallet_ajuna_affiliates::Config<AffiliatesHeroJamInstance> for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type Currency = Balances;
-	type WhitelistKey = ();
-	type AccountManager = CasinoJamSage;
-	type RuleIdentifier = HeroJamRuleIdentifier;
-	type AffiliateMaxLevel = AffiliateMaxLevel;
-	type UnlockParameters = ();
-	type AffiliatesUnlockRules = MockAffiliateUnlockRules;
-	type WeightInfo = ();
-	#[cfg(feature = "runtime-benchmarks")]
-	type BenchmarkHelper = AffiliatesHeroJamBenchmarkHelper;
-}
-
-parameter_types! {
-	pub const TournamentPalletId1: PalletId = PalletId(*b"aj/trmt1");
-	pub const MinimumTournamentPhaseDuration: BlockNumber = 100;
-}
-
-pub type CasinoTournamentCategoryId = u32;
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Encode, Decode, MaxEncodedLen, TypeInfo)]
-pub struct CasinoJamEntityRanker;
-
-impl EntityRank for CasinoJamEntityRanker {
-	type EntityId = AssetId;
-	type Entity = CasinoJamAsset;
-
-	fn can_rank(&self, _: (&Self::EntityId, &Self::Entity)) -> bool {
-		true
-	}
-
-	fn rank_against(
-		&self,
-		_: (&Self::EntityId, &Self::Entity),
-		_: (&Self::EntityId, &Self::Entity),
-	) -> Ordering {
-		Ordering::Equal
-	}
-}
-
-#[cfg(feature = "runtime-benchmarks")]
-pub struct MockTournamentBenchmarkHelper;
-
-#[cfg(feature = "runtime-benchmarks")]
-impl
-	pallet_ajuna_tournament::BenchmarkHelper<
-		CasinoTournamentCategoryId,
-		BlockNumberFor<Runtime>,
-		Balance,
-		CasinoJamEntityRanker,
-		AccountId,
-		AssetId,
-		CasinoJamAsset,
-	> for MockTournamentBenchmarkHelper
-{
-	fn create_category_id(id: u32) -> CasinoTournamentCategoryId {
-		id
-	}
-
-	fn create_default_tournament_config() -> pallet_ajuna_tournament::TournamentConfig<
-		BlockNumberFor<Runtime>,
-		Balance,
-		CasinoJamEntityRanker,
-	> {
-		pallet_ajuna_tournament::TournamentConfig {
-			start: 0_u32,
-			active_end: 100_u32,
-			claim_end: 200_u32,
-			initial_reward: None,
-			max_reward: None,
-			take_fee_percentage: None,
-			reward_distribution: Default::default(),
-			golden_duck_config: Default::default(),
-			max_players: 0,
-			ranker: CasinoJamEntityRanker,
-		}
-	}
-
-	fn create_entities(_: AccountId, _: u32) -> Vec<(AssetId, CasinoJamAsset)> {
-		vec![]
-	}
-}
-
-type TournamentHeroJamInstance = pallet_ajuna_tournament::Instance1;
-impl pallet_ajuna_tournament::Config<TournamentHeroJamInstance> for Runtime {
-	type PalletId = TournamentPalletId1;
-	type RuntimeEvent = RuntimeEvent;
-	type Currency = Balances;
-	type TournamentCategoryId = CasinoTournamentCategoryId;
-	type EntityId = AssetId;
-	type RankedEntity = CasinoJamAsset;
-	type EntityRanker = CasinoJamEntityRanker;
-	type AccountManager = CasinoJamSage;
-	type AssetManager = CasinoJamSage;
-	type MinimumTournamentPhaseDuration = MinimumTournamentPhaseDuration;
-	type WeightInfo = ();
-	#[cfg(feature = "runtime-benchmarks")]
-	type BenchmarkHelper = MockTournamentBenchmarkHelper;
 }
 
 pub const fn deposit(items: u32, bytes: u32) -> Balance {
